requirements.txt
# Image Analyzer Bot - Dependencies

# Web Framework
flask>=2.3.0

# Image Processing
Pillow>=10.0.0

# Google Generative AI
google-generativeai>=0.3.0

# Environment Variables
python-dotenv>=1.0.0

# Security
werkzeug>=2.3.0
